refactor: edit comment component and article component
This commit refactors the comment component and article component. It updates the HTML template of the comment component to improve the styling of the like button. It also updates the TypeScript file of the comment component to handle the like functionality correctly. In the article component, it initializes the totalComments variable properly and updates it when retrieving the article. Additionally, it fixes a bug in the countComments function.